Register numeric parameters correctly and bind binary data values
Why: live regressions showed the legacy WithPrecision registration populates
length-based mappings rather than the parameterized map, leaving DECIMAL at
its default shape. Binary round-trip tests also exposed a shared Insert path
that rejected byte arrays before the database driver could receive them.

Changes: register explicit precision/scale through WithParameters, matching
existing SQL Server/PostgreSQL dialects. Bind byte arrays as DbType.Binary and
bytes as DbType.Byte without changing parameter values or public signatures.

Validation: local build passes. Existing live regressions verify requested
numeric shapes and persisted binary bytes; the ASE byte test now also binds an
actual CLR byte. Full database CI runs on this commit.

+++ src/Migrator.Tests/Providers/Live/LiveMetadataRegressionTests.cs
@@ -125,7 +125,7 @@ public class LiveMetadataRegressionTests
         var column = f.Provider.GetColumns("bytes").Single();
         Assert.That(column.Type, Is.EqualTo(DbType.Byte));
         f.Provider.AddTable("copied_bytes", column);
-        f.Provider.Insert("copied_bytes", ["byte_value"], [255]);
+        f.Provider.Insert("copied_bytes", ["byte_value"], [(byte)255]);
         Assert.That(Convert.ToInt32(f.Provider.ExecuteScalar("SELECT byte_value FROM copied_bytes")), Is.EqualTo(255));
     });
